wip: latest progress
