+++ services/finance/src/app/api/context.tsx
@@ -1,11 +1,12 @@
-import {FC, useRef} from 'react';
+import { useRef } from 'react';
 import { createContext, useMemo } from 'react';
 
+import { ApiQueryReactions } from "./query-reactions";
 import { WalletsMockRESTApiClient } from "@feature/wallet";
-import { TransactionMockRESTApiClient } from "@feature/transaction";
+import {TransactionMockRESTApiClient, TransactionTestRESTApiClient} from "@feature/transaction";
+import type { FC } from 'react';
 import type { IWalletsRESTApiClient } from "@feature/wallet";
 import type { ITransactionsRESTApiClient } from "@feature/transaction";
-import { ApiQueryReactions } from "./query-reactions";
 
 
 interface ApiContextType {
@@ -27,17 +28,17 @@ const ApiProvider: FC<ApiProviderProps> = ({ children }) => {
 	const walletsClient = useRef(new WalletsMockRESTApiClient("wallets"));
 	const transactionsClient = useRef(new TransactionMockRESTApiClient("transactions", "wallets"));
 
-  const contextValue = useMemo<ApiContextType>(() => ({
-    walletsClients: { rest: walletsClient.current },
-		transactionsClients: { rest: transactionsClient.current }
-  }), []);
+	const contextValue = useMemo<ApiContextType>(() => ({
+		walletsClients: {rest: walletsClient.current},
+		transactionsClients: {rest: transactionsClient.current}
+	}), []);
 
-  return (
-    <ApiContext value={contextValue}>
-			<ApiQueryReactions />
-      {children}
-    </ApiContext>
-  );
+	return (
+		<ApiContext value={contextValue}>
+			<ApiQueryReactions/>
+			{children}
+		</ApiContext>
+	);
 };
 
 ApiProvider.displayName = 'ApiProvider';
